style: adjust components directories
