+++ src/app/pre-register/page.tsx
@@ -1,0 +1,12 @@
+import PreRegisterPage from "@/components/Pages/FormPages/PreRegisterPage/PreRegisterPage";
+
+
+const PreRegister = () => {
+    return(
+        <main>
+            <PreRegisterPage/>
+        </main>
+    )
+}
+
+export default PreRegister;
