+++ src/components/Pages/FormPages/PreRegisterPage/PreRegisterPage.tsx
@@ -1,0 +1,16 @@
+import FormPage from "@/components/FormPage/FormPage";
+import FormLayout from "../FormLayout/FormLayout";
+
+const PreRegisterPage = () => {
+    return(
+        <FormPage>
+            <FormLayout>
+                <form>
+                    
+                </form>
+            </FormLayout>
+        </FormPage>
+    )
+}
+
+export default PreRegisterPage;
