+++ src/components/Pages/Home/About/About.tsx
@@ -1,10 +1,10 @@
 import Container from "@/components/Container/Container";
 import Link from "next/link";
 import { PiArrowUpLight } from "react-icons/pi";
-import logoIcon from "../../../assets/logoIcon.svg";
+import logoIcon from "../../../../assets/logoIcon.svg";
 import Image from "next/image";
-import swiftStockUI from "../../../assets/swifStockUI.jpg";
-import swiftStockMobileUI from "../../../assets/swifStockMobileUI.jpg";
+import swiftStockUI from "../../../../assets/swifStockUI.jpg";
+import swiftStockMobileUI from "../../../../assets/swifStockMobileUI.jpg";
 
 
 const About = () => {
